server/requirements.txt
fastapi
uvicorn[standard]
vosk
requests
websockets
sounddevice
numpy
python-dotenv
pydantic
python-multipart  # For handling file uploads
soundfile  # For reading audio files
deepseek  # Assuming this is a Python package; replace with the correct package name if different
langchain  # For structuring prompts and text generation
sqlalchemy
aiosqlite  # Para soporte async con SQLite
openai
python-dotenv  # Para manejar las variables de entorno  